Octokit v0.0.1 Octokit.Repository

Represents a GitHub repository record.

Types

repo :: String.t | t

t :: %Octokit.Repository{created_at: term, default_branch: term, description: term, fork: term, forks_count: term, full_name: term, has_downloads: term, has_issues: term, has_pages: term, has_wiki: term, homepage: term, id: term, language: term, name: term, open_issues_count: term, organization: term, owner: term, parent: term, permissions: term, private: term, pushed_at: term, size: term, source: term, stargazers_count: term, subscribers_count: term, updated_at: term, watchers_count: term}

user :: String.t | Octokit.User.t

Functions

new(user \\ nil, repo)

Specs

new(user | nil, repo) :: t

Creates a new Repository structure with the bare minimum information.

You can call update/2 to fill the structure with the latest information from the GitHub database.

parse(body)

Specs

parse(String.t) :: t

Parses the JSON body of a GitHub API response to construct a Repository structure.

update(client, repo)

Specs

update(Octokit.Client.t, repo) :: t

Updates with the latest information from the GitHub database.